cmd/hiveview: remove unused library code
